+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/BukkitCore.java
@@ -2,11 +2,14 @@ package com.github.games647.fastlogin.bukkit;
 
 import com.github.games647.fastlogin.core.FastLoginCore;
 import com.google.common.base.Charsets;
+import com.google.common.cache.CacheLoader;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 
 import java.io.File;
 import java.io.InputStreamReader;
+import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
 
 import org.bukkit.ChatColor;
@@ -14,9 +17,30 @@ import org.bukkit.configuration.file.YamlConfiguration;
 
 public class BukkitCore extends FastLoginCore {
 
+    public static <K, V> ConcurrentMap<K, V> buildCache(int minutes, int maxSize) {
+        CompatibleCacheBuilder<Object, Object> builder = CompatibleCacheBuilder.newBuilder();
+
+        if (minutes > 0) {
+            builder.expireAfterWrite(minutes, TimeUnit.MINUTES);
+        }
+
+        if (maxSize > 0) {
+            builder.maximumSize(maxSize);
+        }
+
+        return builder.build(new CacheLoader<K, V>() {
+            @Override
+            public V load(K key) throws Exception {
+                throw new UnsupportedOperationException("Not supported yet.");
+            }
+        });
+    }
+
     private final FastLoginBukkit plugin;
 
     public BukkitCore(FastLoginBukkit plugin) {
+        super(BukkitCore.<String, Object>buildCache(5, 0));
+
         this.plugin = plugin;
     }