+++ bungee/src/main/java/com/github/games647/fastlogin/bungee/BungeeCore.java
@@ -1,6 +1,7 @@
 package com.github.games647.fastlogin.bungee;
 
 import com.github.games647.fastlogin.core.FastLoginCore;
+import com.google.common.cache.CacheBuilder;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 
 import java.io.File;
@@ -8,10 +9,11 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.util.concurrent.ThreadFactory;
+import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import net.md_5.bungee.api.ChatColor;
 
+import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.scheduler.GroupedThreadFactory;
 import net.md_5.bungee.config.Configuration;
 import net.md_5.bungee.config.ConfigurationProvider;
@@ -22,6 +24,8 @@ public class BungeeCore extends FastLoginCore {
     private final FastLoginBungee plugin;
 
     public BungeeCore(FastLoginBungee plugin) {
+        super(CacheBuilder.newBuilder().expireAfterWrite(5, TimeUnit.MINUTES).<String, Object>build().asMap());
+
         this.plugin = plugin;
     }